Fix null bug on Java side!
Whoa, it's been here the whole time? If you set a field to null
that has a non-null default value, it would import as the default value.

I came in looking to see if I could make the importers recover if they
encounter an unknown class. I think I'd need to change the stream
protocol again. If I'm going to do that, it gets more complicated to
support backwards compatibility. And really, I need to just settle it
out.

So I'm currently mulling exporting 2.0, which could import fully
backwards compatibly by referencing the legacy classes, but I'd
copy-and-change for the new stuff. There are improvements to be made
throughout.

This may also help with the coaxer I need to write on the C# side,
because I'm rethinking how types are transmitted... maybe we can make
our lives much easier. Currently exporting has a bootstrap "char" type
but no built-in "string" type. Definitely on the list to fix.

So I'll start noodling with that on a branch.