Use 0 to identify fieldIds.
It was starting at 1. More zeros is better, it will probably compress
better.

Plus, in exporting 2.0, I can use a List for holding various mappings
on import, rather than Maps, if we know we'll always see things come
back in an increasing order starting at 0. There's no point in making
that optimization now as old exports might start at 1 and the extra
complexity for that is super not worth it.