Stop trying to export weird collection types.
This is a potentially breaking change with old code...

If you expect to import a LinkedList or something, you're going to be
screwed.

A while back I changed it so that it stopped trying to reflect upon
collection types and just write their elements in a standard way.
I even avoid getting "outers" from ImmutableCollections...

But I didn't change this, and now I would like to.

This is something along the lines of how collections will be treated in
export 2.0, although in that the collection types will be bootstrap
types and we'll never send a string like "java.util.ArrayList".