Remove loggings; rearrange tests; tighten screws.
Everything looks aok.
Who sends buffers? Now at least we pare the instanceof-ing.
Move the enum check to the top.

I want to remove the allowance for ArgumentMap in here and thought
to make it Exportable, but that would make it write all its fields
and not be detected as a Collection. All hell breaks loose because
the collection support is pretty haphazard actually. Stuff to fix
in 2.0.

Instead of warning about unknown collections, just send ArrayList.
The field in the class should probably be Collection or ICollection...

Also I learned that arrays just write their name. So any type
transformations of their component type get bypassed. It would
be pretty odd if someone took a specialized enum instance and made an
array of its class...