Flattening fix, part 2.
This is slightly psychotic but it seems to work.
The DependencyGatherer now operates in two different modes-- first
it gathers dependencies so that we know the order in which to flatten,
then when we inform it about newly flattened configs, it examines
their config references so that it can remember types for the
"superflat" part.

I guess I could keep it like this, but probably this should be made
less confusing by... wait for it... making two different classes to do
the two different things!