Allow a mechanism for capturing exceptions during init.
So, I don't know why, but there are a bunch of parsing problems that
just cause the config manager to log the booch but proceed full speed
ahead anyway. It will even fully init even if a group shat the bed
and is missing.

Perhaps it works like that so that the config editor can come up
even if there are problems?

So: I didn't want to rock that boat, let's add a Consumer for
exceptions. I guess we're still on 1.6 in here so this is a bit crap...