Start writing strings as "unmodified" UTF-8.
Can I just scream at Java for a sec for this whole thing?
I guess the thinking was to avoid putting a NULL byte into
the byte[] so that C-style code can use the same byte[] safely
as a string, or something.

Stupid idea. It's not like we're preventing the byte 0 from
getting written to a DataInputStream. There will be plenty of
zeros. Hell, there's a good chance that the first byte of the
two byte length prefix is 0.

Anyway, let's just start writing unmodified UTF-8. We'll continue
using the built-in readUTF to read modified UTF-8, but that's OK.
Writing "unmodified" is the universal write, reading "modified"
is the universal read.