Fix for duplicate editable props...
The "background color" shows up twice in the prefs for the config
editor.

I dug around, and it's a "bridge method". I don't know why it's a bridge
method... usually such things are used when there's a covariant return
type or there are generics with a subclass constraining the generic
bound...

Maybe this is something that popped-up from using the new compiler.
In any case: ignore bridge methods. They are created by the compiler and
not something we're interested in.