Added writeVarString() and readVarString().
On the stream, a varint is followed by that number of
bytes that are the UTF-8 encoding of the String.

The REAL UTF-8 encoded bytes, not fucking "modified" UTF-8.

This was going to be the basis for writing all strings in export 2.0..
But for right now I'm using it for something else.