Also omit tracking the standard empty/immutable Collections.
Switched to an identity Set and add all the standard empty collections.

Typically I would prefer built-in classes even to guava, but in the past
we've preferred ImmutableList.of() over Collections.emptyList() because
the Collections ones were implemented stupidly. It looks like that's been
fixed up, possibly even in Java 7. We should probably prefer the
built-in ones now.

(As an aside, the non-empty guava Immutable collections still have
 huge advantages over using Java's built-in "unmodifiable" wrappers.)