Fix to cope with null fields.
The string handling has never been great with null fields, but this
must have previously worked? The default value has been null for
years and the string editor hasn't recently changed either. Java 8?

It continues to be not great: this will denull a field as soon as
the user accidentally presses a key in it.