Provide an affordance to reality.
Looking at actual configs, often a shape is put into a compound shape
purely to get a translation. The compound shape adds a bunch of stupid
overhead for these cases.

Add a new Transformed shape type that merely wraps another in a
translation.

Now that we have this, I could also convert Compound so that it uses
regular ShapeConfigs, and those are Transformed as needed.
However that is much rarer, that some shapes don't need translation
but do need compounding, and we save less by that conversion.

Also added an Exporter.Replacer to do this conversion, as an enum
singleton, squirrelled away in here.