Create a simple shape if a compound has only 1 sub shape.
Maybe I should have just done this in the first place.

I'm actually sorting out the shape conversion somewhat... but
there were times when someone wanted to parameterize a shape and
a transform and so they used the compound. Argh. Now we have two
different things for configuring those together...