Enhance ability to coerce editor values.
- Utilize a "Convertible" interface to ask the
old value to convert to the new...
- Also use the existing Constructor choosing functionality
that can be placed on the new type. This is nice but doesn't
have the ability to reject the conversion. A constructor can't
return null.

It checks the Convertible first, and I can enhance this as needed
to adapt to reality if #2 turns out to be a problem.