Only try copying properties when we don't coerce.
This works a bit better but can still be WEIRD.

Let's say you have a HorizontalGroup with a single Label in it.
The HorizontalGroup has tag "group" and the label has tag "label".

You change the type of the component from Container to Label and
the label moves-up as you probably want it to. It retains the "label"
tag.

You change the type of the component back to Container and it
restores the old value from the type cache, but copies properties over
and changes the tag of the Container to "label". This is weird!