Deprecate Coercible. Something better will come.
Automatically coercing the type is a bit weird when the
'container' type has many of the same properties as the type
being put into it. I tried to cope with that by not copying properties
when that happened, but if you switched to a container type
and then back, and then back again, it would copy properties..

There need to be an explicit editor user action to 'compound' and
'extract' elements.