Groupable, a replacement for Coercible. Checkpoint.
Implement it on ActionConfig, at least some of the way.

Rather than just automatically coercing between container and contained
types, we add explicit controls for grouping, ungrouping, and
regrouping.

I'm trying to balance the API between easy to add to an existing class
and easy to use for the user editing. Different containers may accept
different ranges for the number of elements they accept, and it might
be cool to have each type "publish" their range. But that's adding
another method to the interface, and it's wacky I'd have to instantiate
a dummy prototype value to call the method, but it's somewhat speaking
for its entire class? Always weird. So rather, you just try cramming
values in, and it just succeeds or fails.

It's still suboptimal. What if there is a particular container type that
accepts exactly 3 sub-elements? You wouldn't ever see it on the list
for regrouping unless you happened to have 3 elements. Having it publish
its range could mean that we could show a container type in the list
as disabled, perhaps indicating the valid range.

Also some container types accept infinite elements with a special
optional one. For example ActionConfig.Case, which has an optional
default action. I think I'll always slurp-in the last action when
grouping as the default, and if that's not what you want you should
first add an empty action to the end, convert to Case, and then remove
that default.