Don't need to instantiate to determine if Groupable.
Also: was going to prevent regrouping into the same container,
but that could result in showing the button but not being able to use
it. Maybe if I had done the thing with the ranges, and we showed
containers that were out of range, that would make sense, because you
could see your options.