Convert editor/undo diff handling to use properties.
It should have been properties all along, rather than 'fields' per
DeepUtil. I was just thinking about it this morning, and wondering
if I should take another look to try to convert it to the right thing...
But then it was discovered why there are "null" diffs sometimes: when
the comment changes. That's only a property and not backed by a field!
So I dug in again and it was very straightforward. Trying this out for
a while...