Do property change detection privately.
I'm sorry but what the fuck? Introspector does not even handle arrays.
Yet, fundamental to the system are "properties" that are only understood
in certain circumstances. For ConfigReferences, I kinda get it. I was
making excuses. The properties of a ConfigReference are the args and
the types of those properties may be different from the actual type
of a particular argument, plus you want to know about all potential args
not just the ones specified. So, to truly resolve all the properties
of a ConfigReference you need *context* to access the config being
referenced.. But arrays need no such context. Why don't arrays have
read-only properties for length and a property for each index?

I am not trying to get in and rewrite all the property stuff, and break
who-knows-what, so I moved my 'lineage key' stuff locally and handle arrays
and config references myself. But, wtf.