Fix some more things slightly...
The big thing is that entrySet will now not call an entry
a match if it only matches the key.

So I went in to also make containsKey() do the right thing, but
ended up also fixing-up the entrySet and importantly, making
the entrySet check for matching values when removing an entry!
It's a little weird because this Map doesn't honor the contract of
equals() and hashcode() really anyway, but I do think that we should
try to do the right thing.

Note that "normal" queries should have the same performance as before,
even if they miss.