Fix problem with SpawnTransient having a non-uniform transform.
Fixing with the smallest possible "blast radius" because I'm not
sure I understand fully... would I possibly want to do this anywhere
that a transform is seemingly promoted? Do I want to check the
rotation of the matrix to make sure the operation is safe?

Hooking-in to the existing method in ArticulatedConfig was the
quickest/easiest thing...