Get clyde compiling for me... but something isn't right.
Trying to cast to Vector<TreeNode> caused a shitstorm of errors, it's as
if it has a view of Vector pre-generics, but also gives me the warnings
about using a rawtype. Neither the rawtype or unchecked cast warnings
are suppressable, that must not be a feature of my new jdk.

So anyway this works and should work for others too, but even I get
multiple warnings about raw types and whatnot, and I can't fix them.