Write down the command needed to update the subtree. Crappy.
Why does everything have to be crappy? Why isn't there a way to just
automatically keep this up-to-date? Is that what a submodule is? I
thought those needed special pull instructions too?

There should be a simple foolproof way to just say "keep it up to date"
or "lock it to revision X". Or at least, I naively think so.