UI Scaling, the simple version. Root scales its windows.
I had something else going, where each Container could have a scale, and
there were new API methods to for localToGlobalX() and globalToLocalX()
and so on. Each container had to translate mouse event coordinates
for children, and a cumulative scale was needed for computing proper
scissor rectangles...

Got it all working on the clyde side, but a little bit ugly because
scaled containers needed to be hacked-in to a few places like
UserInterfaceWindow... things were a bit impure...

But then I discovered that SK has a whole bunch of special-case hacks
and all of those were going to need complicated revisiting.

Then I thought of this way: Scale at root, and the "absolute" and
"global" coordinates are basically already scaled. Only when we need to
interact at the level of the Renderer (like when adjusting the scissor)
do we need to know about this global scale.

It means: I can't have a scaled UI containing a container that is
inverse-scaled from the outer UI, which then contains a slider that lets
you ajust the UI scaling in a stable way.

Oh well.