Tweak ConfigSearcher to spend up to 50ms checking per queue run.
In the bare config editor it would run quickly but if you triggered it
from the game or one of the visual tools then it would do 1 check per
frame which is not reasonable.

May tweak this. Could see tracking how many are running statically and
dividing the maxtime accordingly but I'm sure this'll be plenty fine.