Add ButtonGroup2. See notes.
By default a ButtonGroup is in "alwaysSelect" mode.

When not, it would not dispatch a null unselect.

Also if you added a new selected button to a group with no selection, it
should become the selected button. It would, but wouldn't notify anyone
about it. Fix that too.

Then: Reimplemented the whole thing.

Tracking selected by searching the list every time is somewhat bananas,
adding buttons to the group shouldn't be O(n^2).

- Always dispatch unselection events. I don't care if the group is
"alwaysSelect", if you call removeAll() while still listening to it
then you *will be told* that the selection is now null. Cope.

... so after fixing all these little things I got bogged down in strange
behaviors in a few old places. Cutting and running. I can't spend a day
debugging old UIs. So now I have a new ButtonGroup going forward. The
name is unfortunate. Better suggestions welcome. If this had been an
interface with a factory all along, bla bla bla, but for now I guess I
plan on just converting things to use 2 eventually and then renaming it
all.