These apply to all importing; undo some of the rearrangements.
- Some properties of the config are only used during import, I was
  calling these "import properties".
- Some properties are used during import and at runtime.
- Some properties of the config are only used in-game.

Turns out, some of the properties in group 1 depend on those in group 2,
so the separations are not so clean.

Someday we'll move all the comments into tooltips on the config
components, so that you could see oh hey it says that
"ignoreRootTransforms" means something more like "clear all the
children of the root when we clear things.