Finish futzing influence flags, ffs.
For now, leave as a resource editor tool.

I am less confident than I was before. Unhooked from validation
completely, it's likely to be more noise than signal. Retained a menu
item in the resource editor to suggest flags. I don't know if this is
good, but it's a rabbit hole I went down...

I thought this would be quick, but now it is the loathsome combination
of a bunch of code that may be projecting too much confidence. I still
don't know if the influence flag config exists because it was decided to
offload this sort of thing onto the configuring human rather than
determine at runtime, or if it's something that's essentially
unknowable, and we declare influence flags for attachments that may be
added at runtime.

Known: omits wrapped model configs... I don't think they matter for my
purposes and it's time to move-on from this rabbit hole.