ModelViewer and InterfaceTester won't transfer arguments...
Disable some somewhat surprising behavior. Normally if you are editing a
config reference and change to refer to a new config, the editor will
try to "transfer" compatible arguments to the new config. This is
normally very useful. But inside these tools you don't realize you're
editing an ad-hoc config reference, you think you're viewing different
files and suddenly ultra confused when arguments copy-over.

Anyway we can try it like this for a while and see if it bugs us.