Fix clicking scrollbar "well" with a snapping bar...
Scroll max of scroll increment/extent if scrollbar track clicked.
Sigh. Maybe I should just change it by the "scroll increment".
The problem is that the default scroll increment is half a page, which
I don't like for the well, but that's how the buttons on the end work.
Just roll with this for now...