Scene Editor: new icons, fixed soundEnabled not applying on startup
This changes how icons are applied. All icons used to have a solid background and _select variants. Now, they are all transparent, and only the toolbar icons (markers, lights, fog, sound, camera) have alternate _select images, to denote the on/off state.

Code was changed to accommodate that.

Once we can ship flatlaf with clyde, I'll take another pass with this.

Note: the soundEnabled scoped var is now technically obsolete - we're controlling the gain directly through sound manager. It should be safe to rip it out completely, but that's a significant effort to not screw anything up in the process.