Woke up with a wild hair to fix ButtonGroup2...
To catch up, the problem is that the old group could be in a state where
no button was selected but it would never dispatch this fact. Various
code now depends on this and if it were changed would blow up with a NPE
when it did something like clear all the buttons and set new ones.

Burn the bridges: Switch over, add a deprecated method to minorly adapt
for old code, but this will absolutely break compilation...

And there will be places that compile fine that *need checking*.

because otherwise I'd have to @Deprecate the no-args constructor, and
that felt like a bridge too far. So instead my solution is fire.

Chop big, and I'll visit my downstream uses...